requirements.txt
django==4.2
psycopg2-binary
djangorestframework
PyJWT
djangorestframework-simplejwt